requirements.txt
asgiref==3.4.1
certifi==2021.10.8
dj-rest-auth==2.2.1
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
psycopg2==2.9.2
PyJWT==2.3.0
pytz==2021.3
sentry-sdk==1.5.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7